+++ packages/apps/wallet-sdk-example/vite.config.d.ts
@@ -1,0 +1,4 @@
+import { UserConfig } from 'vite';
+export declare const config: UserConfig;
+declare const _default: UserConfig;
+export default _default;