+++ packages/e2e/e2e-base/src/constants/coin.constants.ts
@@ -1,1 +1,2 @@
-export const coinModuleHash = 'wOTjNC3gtOAjqgCY8S9hQ-LBiwcPUE7j4iBDE0TmdJo';
+// export const coinModuleHash = 'wOTjNC3gtOAjqgCY8S9hQ-LBiwcPUE7j4iBDE0TmdJo';
+export const coinModuleHash = '3iIBQdJnst44Z2ZgXoHPkAauybJ0h85l_en_SGHNibE';