+++ packages/libs/client-utils/src/integration-tests/marmalade-collection-policy.int.test.ts
@@ -1,5 +1,5 @@
-import type { ChainId } from '@kadena/client';
-import { createSignWithKeypair } from '@kadena/client';
+import type { ChainId, PactErrorCode } from '@kadena/client';
+import { createSignWithKeypair, getPactErrorCode } from '@kadena/client';
 import { PactNumber } from '@kadena/pactjs';
 import type { IPactInt } from '@kadena/types';
 import { describe, expect, it } from 'vitest';
@@ -304,9 +304,8 @@ describe('mintToken', () => {
       config,
     );
 
-    await expect(() => task.execute()).rejects.toThrowError(
-      new Error('with-read: row not found: non-existing-token'),
-    );
+    const res = await task.execute().catch((err) => getPactErrorCode(err));
+    expect(res).toBe('RECORD_NOT_FOUND' as PactErrorCode);
   });
 });
 
@@ -331,7 +330,8 @@ describe('getCollection', () => {
       },
     });
   });
-  it('should throw an error if token does not exist', async () => {
+
+  it('should throw an error if collection does not exist', async () => {
     const nonExistingTokenId = 'non-existing-collection';
     const task = getCollection({
       collectionId: nonExistingTokenId,
@@ -340,9 +340,8 @@ describe('getCollection', () => {
       host: config.host,
     });
 
-    await expect(() => Promise.resolve(task)).rejects.toThrowError(
-      new Error(`read: row not found: non-existing-collection`),
-    );
+    const res = await task.catch((err) => getPactErrorCode(err));
+    expect(res).toBe('RECORD_NOT_FOUND' as PactErrorCode);
   });
 });
 
@@ -369,8 +368,7 @@ describe('getCollectionToken', () => {
       host: config.host,
     });
 
-    await expect(() => Promise.resolve(task)).rejects.toThrowError(
-      new Error(`read: row not found: non-existing-token`),
-    );
+    const res = await task.catch((err) => getPactErrorCode(err));
+    expect(res).toBe('RECORD_NOT_FOUND' as PactErrorCode);
   });
 });