+++ packages/apps/dev-wallet-example/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @kadena/dev-wallet-example
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies \[67ee479]
+  - @kadena/client\@1.17.0
+  - @kadena/client-utils\@0.12.1
+
 ## 0.2.2
 
 ### Patch Changes