+++ packages/libs/client-examples/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @kadena/client-examples
 
+## 0.1.12
+
+### Patch Changes
+
+- Updated dependencies \[67ee479]
+  - @kadena/chainweb-node-client\@0.9.0
+  - @kadena/client\@1.17.0
+  - @kadena/client-utils\@0.12.1
+
 ## 0.1.11
 
 ### Patch Changes